made all the getColor routines accept float values
git-svn-id: svn://orca.physics.unc.edu/Orca@10569 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/BamDectector/BamDetectorView.m
@@ -80,7 +80,7 @@
 	[path stroke];
 }
 
-- (NSColor*) getColorForSet:(int)setIndex value:(unsigned long)aValue
+- (NSColor*) getColorForSet:(int)setIndex value:(float)aValue
 {
 	return [prespecColorScale getColorForValue:aValue];
 }
